Added sync between all open Canvas pages
Items that are hidden or checked will now send that update to all other
Canvas pages open at that time, telling them to update that item.
Currently, only hidden and checked statuses sync; the State system
needs improvements to be able to have the same done to it.

This required an overhaul of the Message system, which became its own
module; and to the initialization, which became a new class, Requests.

Also exposed classes in main.ts to the global scope in the content
script via webpack. This only happens in development mode for debugging.