Content script now gets updated access token
Added a listener for storage area changes in order to tell all canvas
pgaes to update their access tokens and try to run the initialization
flow again with the new token. Also moved some parts of the
background script into utils/vars for more global access.

Note that this feature still needs more testing; the original code was
not designed with the ability to re-run everything in mind, so some
things still break when re-running.