Moved Page class into UI module
The UI module now include the instance of PAGE as its default export,
allowing it to be imported alongside the global import.

The eventual plan with this is to have the UI module be the only one
that imports jQuery, or to have a new module to act as an interface to
jQuery.

Also reordered one of the classes in the objects module so that enums
and classes are grouped properly.