Improved custom data flow
The custom data flow has been split into several functions to allow for
calling each part only on the page it requires. This was a needed to fix
custom navigation tab positions not being queried for unless on the
modules or grades page. Note that this has increased the number of API
requests, so this may become an issue down the line.

Also deprecated the CanvasAPI.CustomData interface and added the
deprecation rule to tslint to support that change.