Fixed fatal bug in custom SCSS loader
The custom SCSS loader was failing due to the addition of a few values
in vars.ts that contains curly braces; those are now filtered out from
the loader side of things.

Also improved the error reporting of the custom loader - previously, it
was not printing any errors because the callback was receiving a string
instead of an Error object.