Added better custom course color use
Custom course colors were only being saved and applied for favorited
courses for use in the sidebar. Now, colors are stored separately and
will properly load when on the page of a non-favorited course.

Also improved the Data object to create all Map instances in their
declarations instead of in the constructor.