Started custom tab position overhaul
The backend system for custom tab positions is complete, including
updating positions on the page and storing custom positions in the
custom data API.

Also, an addition was made to inject code into the page's global scope
to override some of ECPI's JS functions that were problematic; namely
the function that was removing the "syllabus" anchor element but leaving
the empty list item element. That particular function has been
re-implemented to work better for the custom tab position system.